递归返回树形菜单，实现菜单管理，实现为角色分配菜单

+++ spzx-manager/src/main/java/com/cc/spzx/manager/utils/MenuHelper.java
@@ -25,7 +25,7 @@ public class MenuHelper {
         sysMenu.setChildren(new ArrayList<SysMenu>());
 
         for(SysMenu it : list) {
-            if(it.getParentId().longValue() == sysMenu.getId()) {
+            if(it.getParentId().longValue() == sysMenu.getId().longValue()) {
                 sysMenu.getChildren().add(dfs(it, list));
             }
         }
